+++ src/main/java/cd/go/contrib/elasticagents/openstack/executors/GetProfileMetadataExecutor.java
@@ -34,6 +34,7 @@ public class GetProfileMetadataExecutor implements RequestExecutor {
     public static final Metadata OPENSTACK_NETWORK_ID = new Metadata(Constants.OPENSTACK_NETWORK_ID_ARGS, false, false);
     public static final Metadata OPENSTACK_SECURITY_GROUP = new Metadata(Constants.OPENSTACK_SECURITY_GROUP, false, false);
     public static final Metadata OPENSTACK_KEYPAIR = new Metadata(Constants.OPENSTACK_KEYPAIR, false, false);
+    public static final Metadata OPENSTACK_MAX_INSTANCE_LIMIT = new Metadata(Constants.OPENSTACK_MAX_INSTANCE_LIMIT, false, false);
     public static final Metadata OPENSTACK_USERDATA = new Metadata(Constants.OPENSTACK_USERDATA_ARGS, false, false);
 
     public static final List<Metadata> FIELDS = new ArrayList<>();
@@ -44,6 +45,7 @@ public class GetProfileMetadataExecutor implements RequestExecutor {
         FIELDS.add(OPENSTACK_NETWORK_ID);
         FIELDS.add(OPENSTACK_SECURITY_GROUP);
         FIELDS.add(OPENSTACK_KEYPAIR);
+        FIELDS.add(OPENSTACK_MAX_INSTANCE_LIMIT);
         FIELDS.add(OPENSTACK_USERDATA);
     }