+++ src/main/java/cd/go/contrib/elasticagents/openstack/requests/CreateAgentRequest.java
@@ -20,6 +20,7 @@ import cd.go.contrib.elasticagents.openstack.AgentInstances;
 import cd.go.contrib.elasticagents.openstack.PluginRequest;
 import cd.go.contrib.elasticagents.openstack.RequestExecutor;
 import cd.go.contrib.elasticagents.openstack.executors.CreateAgentRequestExecutor;
+import cd.go.contrib.elasticagents.openstack.utils.OpenstackClientWrapper;
 import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -63,6 +64,6 @@ public class CreateAgentRequest {
     }
 
     public RequestExecutor executor(AgentInstances agentInstances, PluginRequest pluginRequest) throws Exception {
-        return new CreateAgentRequestExecutor(this, agentInstances, pluginRequest);
+        return new CreateAgentRequestExecutor(this, agentInstances, pluginRequest, new OpenstackClientWrapper(pluginRequest.getPluginSettings()));
     }
 }